Changed room db schema and updated dao

+++ app/src/main/java/com/gaurav/pokemon/data/model/ApiToken.kt
@@ -8,5 +8,5 @@ import java.io.Serializable
 data class ApiToken(
     @PrimaryKey
     val token: String,
-    val expiresAt: Double
+    val expiresAt: Long
 ) : Serializable
